Importer UI part deux (#1221)
* Continuing work on importer UI...

* Progress updating the importer views.

* Removing refactored create view into separate import specific views.

* Fixing page titles.

* Lots more refactoring to improve understanding.

* Updating RTV test form.

* More updates, adding show views for import resources.

* Updating docblock for PHP format.

* Enabling all aspects of Mute Promotions UI.

* Progress on enabling Email Subscriptions UI.

* Apply suggestions from code review

Recommended changes!


* Refactoring and enabling email subscription imports.

* Cleanup

* Using FileImportController base class

* Removing FileImportController in favor of helper function and Parse Jobs.

* Switching so if either config is disabled, code is skipped.

* Updates and expanding on RTV import controller.

* Adding comment.

+++ app/Services/CustomerIo.php
@@ -171,6 +171,12 @@ class CustomerIo
      */
     public function deleteCustomer(string $id)
     {
+        if (!config('features.customer_io') || !config('features.delete-api')) {
+            info('User ' . $id . ' would have been deleted in Customer.io.');
+
+            return;
+        }
+
         logger('Deleting Customer.io profile', ['user_id' => $id]);
 
         return $this->trackApiClient->delete('customers/' . $id);
@@ -184,7 +190,7 @@ class CustomerIo
      */
     public function suppressCustomer(string $id)
     {
-        if (!config('features.delete-api')) {
+        if (!config('features.customer_io') || !config('features.delete-api')) {
             info('User ' . $id . ' would have been suppressed in Customer.io.');
 
             return;
